backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
openai==1.88.0
langchain==0.3.26
langchain-openai==0.3.27
langchain-chroma==0.2.4
langgraph==0.5.1
langchain-mcp-adapters==0.1.9
fastmcp==2.10.5
pydantic==2.11.7
python-dotenv==1.1.0
requests==2.32.4
RapidFuzz==3.13.0
pypdf==5.7.0

sentence-transformers==2.6.1
torch==2.2.1
torchvision==0.17.1
transformers==4.48.0
numpy==1.26.4